Fixed problem with nested functions and refactor
Fixed problem where nested functions where
considered as normal top level functions and
therefore broke the content of the parent
function

Also did a major package/module refactor.
Moved plenty of classes to make the structure
more logic.
Now, each parser is in its own class, and outputs
a dedicated data object, also stored in the same
module.
Unit tests for the parsers are (or should be)
directly in the same file.