Improved the line parser quite a lot
Now the line parser is able to return an array of
all lines, with each item being a Line object.
Each line object has a line number, a complete
source code, only code, and only comments.
As well as a series of useful functions.

Adapted the rest of the source code to take advantage
of this.