Fixed the get/has/is/set function reviewers
This fixes #21 whereby a function starting with set and
not accepting an argument was outputing a warning, even
when that function was "setup" or "settle" for instance.
Same kind of problem with functions starting with get
has or is and not returning something.

There is now a real (unit tested) parser for function
name prefixes as well as for return statements inside
functions