Big refactor to move to true JS parser
Now, the function and variable parsers are just visitors
on top of a true jsparser (based on pynarcissus).
The line parser is still good old string walking but so
far we still need this cause we want comments too.

This move simplifies the code, removes a great deal of
regexp, and moves the complexity to pynarcissus where it
should be.
Unit tests have therefore decreased a lot at parsers level
cause there's not much point testing that a visitor's methods
get called when the tree is walked.
Integration tests still pass.

The most immediate effect of this move is that it fixes #24