First implementation of the unused class property reviewer
The parser now extract the list of this.... class properties
And then, the 'unused' reviewer checks all lines to see whether
these properties are used or not.
This will typically fail if there are 2 classes in the same file
that have a property named the same.
But hey, that's the first implementation. It will require true
class parsing to be perfect. And the thing is, true class parsing is
hard to achieve knowing all those frameworks out there that have their
own way of declaring classes.

Anyway, this is the first step for #12