Finally fixed the file open path problem
Now, the reviewer takes an optional argument which is the working
dir.
When running cleanjs from the command line, or when running the unit
tests, then this argument is not set, because the normal getcwd()
returns what is needed.
In the case of sublime and appengine website, then this argument is
set to the proper value so that the file can be found.